requirements.txt
streamlit>=1.26.0
ultralytics>=8.0.20
streamlit-webrtc>=0.45.2
opencv-python-headless>=4.5.5.64
pillow>=9.5.0
pandas>=1.5.3
av>=10.0.0
